requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
mfrc522==0.0.7
python-statemachine==2.3.1
requests==2.32.3
RPi.GPIO==0.7.1
RPLCD==1.3.1
smbus2==0.4.3
spidev==3.6
urllib3==2.2.1